Back to separate channels
After consulting LAD and thinking a bit more, I decided to go back to an
internal representation of multiple channels instead of interleaved data. I
also opted for a much simplified API, much more like an NArray (which,
actually, it is a subclass of). Now it is easy to work with channels in-place,
and to get/set from interleaved data as needed.

darcs-hash:20051115225541-28bda-84dbc1490a274d7e37051c9543bdd35c2e81e41e.gz